refactor: separate submodules of `utils`
